requirements.txt
# LangGraph Implementation Agent 依赖

# 核心LangGraph和LangChain依赖
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# Anthropic Claude API
anthropic>=0.34.0

# OpenAI API (可选)
openai>=1.40.0

# 数据处理
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 异步支持
aiohttp>=3.8.0
aiosqlite>=0.19.0

# 输出格式支持
PyYAML>=6.0
markdown>=3.4.0

# 数据库和存储（用于LangGraph检查点）
# sqlite3 是Python内置模块，无需通过pip安装

# 开发和测试依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# 类型检查
mypy>=1.0.0
types-PyYAML>=6.0.0

# 代码格式化
black>=23.0.0
isort>=5.12.0

# 文档生成
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0

# 性能分析
memory-profiler>=0.61.0
psutil>=5.9.0

# 日志
structlog>=23.1.0
rich>=13.0.0

# 环境管理 (必需)
python-dotenv>=1.0.0